PsychLinuxConfiguration: Handle xorg.conf.dir on RaspberryPi OS 12.
RaspberryPiOS 12 can create its own /etc/X11/xorg.conf.dir/ - owned by
user and group root, to store config files specific to RPi4+. In that
case, our detection of "missing xorg.conf.dir" skipped, as the dir was
there, but then group ownership of the dir was by root, not by our group
psychtoolbox, therefore it was not writable by XOrgConfSelector.

Be more clever and handle the case of an existing dir with wrong group
ownership, and in that case, just offer to change group ownership to
psychtoolbox, with suitable write permissions.

-> Should fix failure of XOrgConfSelector() on RaspberryPi OS 12+