[Legacy]XOrgConfCreator(): Only expose 10 bpc mode on capable RaspberryPi's.
Check version of Mesa and of the VideoCore's V3D engine, ie. the version
of VideoCore. We only offer deep color 30 bit / 10 bpc setup if this is
a VideoCore 6 or later (RaspberryPi 4/400/5 and equivalent), driven by
Mesa 23.3.0 or later. Older Mesa versions or hardware, e.g., VideoCore 4
of RPi 1/2/3 would fail X-Server startup.

Note we allow Mesa 23.3.0, although support will only be officially
available in Mesa 23.3.1, but the current Mesa 23.3 staging branch will
expose upcoming Mesa 23.3.1 as Mesa 23.3.0, and no version of RaspberryPi
OS ships with Mesa 23.3.0 by default, so this should be both neccessary
and safe to do for now.