Add new GitHub Action for building a zip and MLTBX file. (#258)
This Action is tag driven, it releases zip and MLTBX files in
GitHub's release section with the main Psychtoolbox folder
of a new tagged release.

- The Psychtoolbox zip file can be downloaded and unzipped by
  users, then one can run `SetupPsychtoolbox` in the new
  uncompressed Psychtoolbox root folder.

- The MLTBX Matlab toolbox file can be downloaded and installed
  via Matlab's Add-On manager, but the user still needs to change
  to the toolbox root folder and manually run `SetupPsychtoolbox`
  to get a functional toolbox add-on.

This is the work of Bensingh Pancras from Mathworks open-source
community support team. It serves as a starting point for further
needed refinements, but serves as a proof of concept baseline for
further work.

Working principle:

The release files will have the file names in the following format
{tagName}.{zip, mltbx}. The Action is set to work with the current
tagging convention of PTB.

  buildZipMLTBX.yml implements the workflow for the Action
  The workflow uses createPsychtoolboxMLTBX.m for creating the MLTBX
  The configurations for the MLTBX packaging are located in Psychtoolbox.prj

The configurations within the PRJ file can be viewed by opening the
file in MATLAB.

Please verify that the metadata in Psychtoolbox.prj is accurate.

"THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES HOWEVER
CAUSED AND ON ANY THEORY OF LIABILITY."

---------

Co-authored-by: Bensingh Pancras <bpancras@vdi-dd1bgl-042.dhcp.mathworks.com>
Co-authored-by: Bensingh Pancras <bpancras@mathworks.com>