GitHub action for zip+MLTBX: Move Psychtoolbox.prj back to root folder.
Apparently the MLTBX builder does not like it if the .prj file
is somewhere else than in the Psychtoolbox-3 project root folder,
as it derives the path of the to-be-included-files base folder
Psychtoolbox from the location of Psychtoolbox.prj.

I don't have the time right now to deal with this elegantly, given
that this is supposed to be vacation day nr. 2 (hahahaha!), so
just move Psychtoolbox.prj back to the root folder.

This new GitHub actions workflow definitely needs more work and
thought in the future, but now is not the future.