EyelinkToolbox: Further fixes to broken legacy functions.
EyelinkDoDriftCorrect(), EyelinkLegacyDoTrackerSetup(), and
EyelinkTargetModeDisplay() had various references to no longer
existing other legacy functions, as the legacy fixes were
incomplete. Fix up names of functions in calls.

Also improve legacy warning() messages and help texts a bit,
to tell users what to do instead of using legacy functions.

This new legacy code was clearly not tested at all,
and if it's not tested, it's broken! Ofc. it is still not tested,
as I lack the needed hardware.