EyelinkDoDriftCorrection(), EyelinkDoTrackerSetup(): Fix broken video support for Octave.
Nested functions are used needlessly. These can not operate on global
variables like eyelinkanimationtarget in Octave as they'd do on Matlab,
so this is totally broken. Clearly this code has not been tested on
Octave, at least not on Octave 5.2 which is the version we must still
support for Eyelink toolbox, as long as we still support Ubuntu 20.04 LTS!

Rewrite handling of global eyelinkanimationtarget to fix this.

Testing of this is very limited, as I lack Eyelink hardware to properly
test, so only the obvious bugs were fixed.