EyelinkToolbox: Improve / fix new audio feedback tone output path.
- Clarify some warning / help messages.
- Improve interop with Snd() in the Snd() fallback path.
- Bug fix in the fallback path for unknown sound spec in EyelinkMakeSound.

I am not convinced that having the Snd() fallback at all is a good
idea, given we now have 3 separate sound output implementations,
depending if el.ppa_pahandle is passed or not, and if
EyelinkUpdateDefaults() is called or not, with three different behaviors.
This may confuse users.

But at least it is probably less buggy, and the design is a different
decision.