PsychCV mex for 64-Bit Octave/Matlab on Ubuntu Linux built.
Ubuntu 20.04.6-LTS build system, Octave 5.2, Matlab R2023b.
This links against libapriltag.so.1 and libapriltag-utils.so.1,
ie. major version 1 of the apriltag library, as shipping with
Ubuntu 20.04-LTS.

Note that Ubuntu 22.04-LTS and later ship with major version 3
of the apriltag library, which is mostly API and ABI compatible,
at least enough for our purposes. However, the so name has changed,
so for using the built PsychCV mex file, one needs to create soft
links from libapriltag.so.3 to libapriltag.so.1 on Ubuntu 22.04-LTS
and later, or the mex file will fail to link and load.

We may drop support for Ubuntu 20.04 for this mex file soon, and
then the library name hack can go away.