PsychCV mex for 64-Bit Octave/Matlab on Ubuntu 22.04 Linux built.
Ubuntu 22.04.4-LTS build system, Octave 6.4, Matlab R2023b.
This links against libapriltag.so.3 and libapriltag-utils.so.3,
ie. major version 3 of the apriltag library, as shipping with
Ubuntu 22.04-LTS and later.

It will not work on Ubuntu 20.04-LTS or any versions older than
Ubuntu 22.04-LTS. That's fine, as 20.04 is considered mostly in
maintenance mode, and the older libapriltag version 1 of 20.04
LTS is inferior anyway, so any serious user of this would use
22.04-LTS and later anyway.