Screen/Linux: Fix idiotic bug in Screen('GetFlipInfo') there since day 1!
If a query for a flip info struct was made, but no new struct was available,
the function errored out, instead of returning an empty [] return argument!

This was not ever noticed due to the way our only test scripts for this
PerceptualVBLSyncTestFlipInfo2 had try-catch-end statement placed just so
that the error was swallowed and never showed up!

Fix this properly by returning an empty [] return argument in this case,
as expected.