MOGL: Fix wrong definition of glPathGlyphIndexRangeNV().
This causes compiler errors recently, so fix it. This due to spec weirdness:

Spec says GLuint baseAndCount[2] in some places, GLuint* baseAndCount in others,
and our parser chose the first definition as GLuint baseAndCount, because it does not
know how to handle this one and only instance ever, where an array is defined.

Fix this up by manual implementation and fixup interface on Matlab side as well.