PsychVulkanCore: Adapt includes to allow use of GStreamer SDK in lieau of Vulkan SDK.
GStreamer 1.24 devel contains all Vulkan and MoltenVK header files recent enough, and
a libMoltenVK.a static link library as universal library, to allow us to avoid installation of
the Vulkan SDK and MoltenVK for building and running PsychVulkanCore + MoltenVK
on macOS. We need some include and defines trickery to make this work though, as the
GStreamer provided headers are not complete and up to date enough.

This may not be a future-proof method to do it, e.g., MoltenVK as bundled with GStreamer
is outdated wrt. upstream, but for the moment this does the trick well enough.