CedrusResponseBox(): Disable any flow control.
It was supposed to use hardware flow control (FlowControl=Hardware),
but that did not work until PTB v3.0.19.6 due to an IOPort parameter
handling bug. Since the bug is fixed, hardware flow control would be
used. This was supposed to make Cedrus to work more robust, but it
apparently had the opposite effect, as the following PTB forum message
suggests:

https://psychtoolbox.discourse.group/t/waiting-forever-for-data-transmission-with-cedrus-rb/5319

Disable flow control explicitely, to revert to pre 3.0.19.6 behaviour,
where Cedrus boxes worked as well as they could.

If this fixes the bug though, that means my original opinion about
the quality of Cedrus hardware - or the lack thereof - would be
true, and the fragile communication was not caused by our IOPort
driver bug, ie. by lack of hardware flow control. Let's see...