PsychCV('AprilDetectMarkers'): Make sure to choose best candidate for each marker.
If the same Apriltag marker id is detected multiple times in an image by
the apriltag library, e.g., due to some false positive detections, we only
return one detection as marker of that id. As opposed to the past, when
we just returned the first detected candidate of a certain id (first come,
first served) instead return the candidate detection with the "highest
quality".

Best candidate detection is defined as the one with the lowest hamming bit
error, and in case of multiple candidates with same hamming bit error the
one with the highest decision margin (aka "score").

Practical testing showed that this greatly improves stability of marker
detection in a real use case.