Screen('OpenMovie'): On Linux, use pipewiresink as audio sink if available.
Instead of leaving audio sink selection to GStreamer, we use pipewiresink
as override sink for the playbin. Why? Because there is an unfixed bug
in PulseAudio 16.x libpulse client library that can cause unwanted lag
in start of movie playback, proportional to the time passed between
Screen('OpenMovie') and Screen('PlayMovie'). The bug was fixed in
PulseAudio 17, but current Debian and Ubuntu do not ship PulseAudio 17
or any backport of the bug fix, as of Ubuntu 24.04.0-LTS.

This bug affects Ubuntu 22.04-LTS and later as of now. Our workaround
uses pipewiresink, which is available on Ubuntu 24.04.0-LTS and later
and directly sends audio to the Pipewire sound server, thereby circumventing
the Pulseaudio client library bug and fixing the problem.

Ubuntu 22.04-LTS can not be fixed this way, and 20.04-LTS was not affected.

See following issue for reference:
https://github.com/Psychtoolbox-3/Psychtoolbox-3/issues/814

Hopefully Fixes #814 for Ubuntu 24.04-LTS and later.