TwoStateQuery(): Add workaround for broken Matlab R2024a on Apple Silicon macOS.
questdlg() is broken on at least R2024a Apple Silicon macOS. Sometimes it works, but most often if one
clicks on a dialog button, Matlab hangs in the spinning beach ball of death. No such problem with Matlab
on Linux, or with Octave on macOS.

Fall back to ugly figure() method for Matlab on macOS ARM for the time being.