Cohorts/HID Utilities for macOS: Fix universal compilation on recent XCode.
These are missing changes to XCode project files and settings which were needed
for the universal build of libHID_Utilities64.a from a release about six months ago.

Most important was the addition of the #define of
__ASSERT_MACROS_DEFINE_VERSIONS_WITHOUT_UNDERSCORES to bring some
legacy macros back. And some changes to build universal static 64-Bit only libs.

Other changes may be not needed, and just change workspace/editor settings etc.,
but I'm not bored enough to untangle which are incidental and cosmetic, and which are
not. This should work on at least XCode 14.