Screen/Linux/X11: Add braces for NetWM reporting to make compiler happier.
Lack of braces should imho not be a problem, and wasn't in the Octave
builds of Screen.mex, but the build for Matlab has a weird problem, in
which the "PTB-INFO: ..." lines about the used flip timestamping mechanism
are not printed! No clue why, looks almost like some compiler bug, but I
accidentally found this today, so fix it.