PsychImaging: Fix HDR with some types of PsychColorCorrection().
Turns out we were missing one argument in HDR path setup, which makes
color correction based on LUT's or gain matrices fail. Ie. because the
lookup texture does not get bound, we only see black. A minor omission
caused this, fix it.

It did work before the fix if at least two color correction tasks are
used at the same time, so the buggy one-colorop optimization via downstream
color correction inside the output formatter failed. Now it always works.

When we are at it, enhance the HDR10-PQ_Shader, so it allows one-pass
color correction in all cases, and thereby not only on Linux, but also
on MS-Windows and macOS. Principle of "least surprise" and all that...