linuxmakeitoctave3: Basic build support for ARM64 and 64-Bit Fedora.
Define preliminary target folder for mex files built for/on 64-Bit ARM.
Not neccessary the final word on this, but will do for current experiments.

Enable basic building on Fedora Linux and spins/remixes for 64-Bit cpu
architectures.

Build WaitSecs to identify as a Wayland mex file when building on a
Wayland system. This mostly for the benefit of LM stats collection about
Linux WSI type.