PerceptualVBLSyncTest[FlipInfo][2]: Add workaround for gpu's withoug hw cursors.
Some 64-Bit ARM platforms display engines or drivers do not support
hardware cursors, so displaying a mouse cursor will enforce fallback
cursor rendering that may impair visual timing significantly, which is
not what we want in these test scripts. As a workaround, HideCursor()
the cursor on 64-Bit ARM for that reason.

When we are at it, convert PerceptualVBLSyncTestFlipInfo.m from DOS
CR+LF to Unix LF line endings to make things less disgusting.