FlipTimingWithRTBoxPhotoDiodeTest: Let audio card select its preferred sample rate.
Some recent sound chips don't like our hard-coded 44100 Hz, so leave it
to the audio system to choose what it can handle. Doesn't matter if sound
sounds weird during this test, as it only fills the purpose of creating
some processing load in the audio system, not to please the ears of the
listener.