Screen/Wayland: Add some hacks to explore how to handle HiDPI/Retina scaling.
These are hacks, and only half of them work halfways, but better than
nothing, and if only to document what doesn't work well so far, so we
don't repeat past mistakes again out of forgetfulness...

Especially the hack in PsychWindowGlueWayland.c was totally useless, and
is therefore coded to be a no-op due to bscale = 1; as other bscale values
do more harm than good.

The hack in PsychScreenGlueWayland.c is implemented in the wrong place,
but shows some promise, just needs a more careful implementation for future
production use.