MultiWindowLockStepTest: Workaround for Windows OpenGL quirk.
After the first AsyncFlipBegin on a Window, no OpenGL context is active, therefore
one either needs to activate one via any drawing command, or GetWindowInfo,
or a glGetError() check in the immediately following AsyncFlipBegin for the 2nd
onscreen window would fail on MS-Windows due to an OpenGL error handling
oddity of MS-Windows. Lets force the OpenGL context on via an otherwise
redundant call to Screen('GetWindowInfo', w2);

No such problems exists on Linux or macOS, where the OpenGL drivers behave
reasonable and glGetError() returns GL_NO_ERROR if no context is bound, as the
spec requires.

This fixes the problem in the 2-Window 10/20 seconds delayed swap case on
Windows-10 for AMD and NVidia graphics. Long term we need a better fix for
this inside Screen's preflip operations path...