managementtools: Add setup function for LexActivator SDK and runtime libs.
The new function downloadlexactivator(alsosdk, cleanupafter); downloads
the zip files with the LexActivator SDK and runtime libraries from our
license management provider Cryptlex. Then it unzips the files and moves
the proper SDK header/include files and Windows import library into place
to allow compiling/linking/building mex files with LexActivatore license
management support. This is done (for local or CI builds) if the parameter
'alsosdk' is set to 1. Otherwise ths SDK is skipped and only the runtime
libraries are installed, e.g., for inclusion into a PTB release zip file.

Downloaded zip files and temporary directories are deleted on success or
failure if 'cleanupafter' is set to 1, otherwise files are kept.

Typical uses:

1. For local build machine setup or setup of CI build, call
   downloadlexactivator(1, 1); or downloadlexactivator(1, 0); to keep
   local backups of the files for backups/documentation etc.

2. For building a zip file with only the Psychtoolbox folder as part of
   an official public PTB release triggered by the maintainer, or a
   MLTBX file, call downloadlexactivator(0, 1); to only put the runtime
   libraries into the to-be-zipped-up Psychtoolbox folder, ie. into the
   Psychtoolbox/PsychBasic/PsychPlugins folder and its Intel64 and ARM64
   subfolders.

This is an initial workable release. The name and location of this
function file may change, or not. Not yet sure about the best place...