Screen/OSX: Skip sync tests and calibrations on Apple Silicon gpu with Vulkan display backend.
With Apple proprietary OpenGL for Apple Silicon graphics on macOS, the sync tests will always fail,
because timing and sync is broken with Apple OpenGL, and the beamposition queries are completely
unsupported due to the PsychtoolboxKernelDriver being unsupportable on Apple DCP.

Save the time and output clutter of the doomed sync tests when the external Vulkan/Metal display
backend is used for display of fullscreen onscreen windows. This backend uses its own presentation
timing and timestamping and may need to implement its own custom synctests in the future.

We leave the doomed tests in place for windowed and/or non-Vulkan mode, so the user gets visible
warning about the certainly broken visual timing and timestamping.