Screen/OSX: Add movie playback support in pixelFormat 11, ie. WCG / HDR.
Add a variant of the WCG/HDR decoding shader which is only selected on macOS and works around
the primitive OpenGL implementation of macOS, which is OpenGL 2.1 only, GLSL 1.2 only, and thereby
misses support for isnan() and switch-case statements, which would require GLSL 1.3 ie. OpenGL 3.0
at a minimum. As opposed to the standard variant of the shader for Linux and Windows, we work
around lack of switch-case by implementing an equivalent if-else ladder, and avoid use of isnan().

This makes the shader - and thereby pixelFormat 11 WCG/HDR/4k+/10bpc+ playback work on macOS,
although likely at a performance hit compared to Linux and Windows.