Screen/OSX: Fix GStreamer videorecording on Apple Silicon.
Since GStreamer 1.20, which is the minimum required GStreamer version for Apple
Silicon macOS, video recording was broken. This also on macOS for Intel Macs, but
there users could work around the problem with a downgrade to GStreamer 1.18.

Now we need a better fix for Apple Silicon. Debugging and reading GStreamer
source code suggests roughly the following problem: When video recording is
used, the element after avfvideosrc exposes caps that allow use of GLMemory as
image backing storage, ie. OpenGL textures, but the plugin can not actually handle
that for some reason and crashes. So we need to try to prevent that, and do so by
adding a capsfilter element behind the avfvideosrc, which limits allowable src pad
caps to standard video/x-raw, disallowing use of GLMemory, thereby preventing
the crash.

This seems to work, as tested with GStreamer 1.24.10 on macOS 14.5 on an Apple
M2 Pro with builtin FaceTime HD camera. VideoRecordingDemo.m now works and
writes a valid movie, with and without live video feed. VideoCaptureDemo.m
continues to work fine.

What still doesn't work on macOS for Apple Silicon is sound recording though.
Enabling that will freeze the video recording after a few seconds of recording.

When we're at it, update VideoRecordingDemo.m to use the default codec for
movie recording, as that now seems to work again.