IsARM(): Add query for true host system architecture for macOS.
Normally IsARM() reports the architecture for which the Octave/Matlab scripting
language runtime - and Psychtoolbox mex files in use - are compiled for, ie. if that
target architecture is ARM or not. However, emulation layers like Rosetta2 on macOS
allow running a Matlab/Octave/Mex file for Intel Macs to run on ARM Macs, and so
the reported IsARM() state may be inappropriate for some decision making.

The new optional parameter IsARM(1) now queries true host architecture, even
in case of Rosetta2 emulation, so we can treat Apple Silicon even when operated
from executables originally targeted at non-ARM Intel hardware.

The detection logic is not yet implemented for Linux (cfe. FEX-Emu or Box64) or
Windows on ARM's emulation layer. That's a todo for the future, but not as
urgent as is for macOS atm.