Switch PsychImaging/PsychVulkan et al. for IsARM(1) to deal with Rosetta2.
Use the new query function to make sure Apple Silicon systems are treated as such,
even when operated from a PTB for IntelMacs running under Octave/Matlab for
IntelMacs.

While this should help those users running Matlab for IntelMacs on their Apple
Silicon, note that PTB has not and will not be carefully tested for proper working
under such configurations. It's more of a step-gap.