PsychLicenseHandling(): Add auto download and install for LM client libs.
Detect absence of required license manager client libraries during 'Setup' and try
to auto-download and install the missing crucial libraries.

This is relevant for non-standard Psychtoolbox installations via GitHub clone or
checkout, download of source tree zip/tgz files, or maybe even when the dreaded
deficient Matlab Add-On manager is used if we are lucky.