downloadlexactivator(): Update to Cryptlex SDK/Runtime v3.31.2.
Apart from minor bug fixes and minor new api, and ARM64 support for Windows,
which we don't use atm., the main improvement is support for system-wide
activations on macOS, which should be quite useful for our users.

Note that we now use Windows dll's built with MSVC2015 (vc14) instead of
MSVC2019 (vc16), because the vc16 folder with MSVC2019 builds was dropped
by Cryptlex for some reason. I hope this won't cause compatibility issues, but
our extended test phase should show that in the coming week.