Screen/Windows: Whenever GStreamer is used, disallow MMCSS scheduling.
Turns out that MMCSS scheduling of the Psychtoolbox main thread, aka Matlab /
Octave main interpreter thread, for improved realtime behaviour wreaks havoc
whenever GStreamer is used for any multi-media tasks like movie playback, movie
writing, video capture or video recording.

Specificially, Octave/Matlab hangs/crashes/aborts if we are in Priority(1) or higher
mode with MMCSS on the main thread active during GStreamer startup, ie.
gst_init_check(). This because GStreamer uses GLib for creating its own many
processing threads and GLib tries to make new created threads inherit the
thread scheduling priority of their parent/creator thread. This mechanism is not
able to cope with the parent being MMCSS scheduled, causing an invalid
parameter error in SetThreadPriority() and then the error handling to kill off the
whole thread with a very unhappy ending. Dealing with non-MMCSS scheduled
parents is not a problem.

So short of fixing GLib for Windows thread creation code and then waiting for
a new GLib release propagating into a new GStreamer release, the best we can
do is to disable MMCSS scheduling of the master thread - and then leaving it
disabled for the remainder of a session - as soon as GStreamer is about the get
initialized. Not a great solution, but better than crashing/hanging on Windows.

In practice this means we have the superior MS-Vista and later MMCSS scheduling
as long as no multi-media functionality is used, and fall back to conventional and
less timing stable WindowsXP style scheduling as soon as a script uses a GStreamer
backed function, the most prominent one being video playback.

Note that at least an admin user could also launch Octave/Matlab as Administrator,
to unlock full Priority(2) support to get MMCSS performance without MMCSS, as
one workaround if one needs to push RT behaviour and use GStreamer functions.

Relevent GLib code that is the culprit starts here:

https://github.com/GNOME/glib/blob/97bd09a04f582a4c87568b3826e39c265efba31b/glib/gthread-win32.c#L479

Bug report by @dcnieho which triggered this diagnostics and fix here:
https://github.com/Psychtoolbox-3/Psychtoolbox-3/issues/857

Hopefully fixes #857 at least it does on my Windows-10 machine as tested with
Octave 7.3 and Matlab R2024a and GStreamer 1.24.10 on AMD Ryzen gpu and
iGPU.