LM: Add auto-activation support for a node via config file.
The LMOpsAllowed.txt file is now also searched in the Psychtoolbox root folder, so admins
can store a file there for global or shared installations, or disc imaged deployment. The
files existence signals consent, as before. In addition, storing a license key in the first
line of the file will cause an attempt at enrolling that key if a valid key isn't enrolled already,
and a license activation if none is activated yet.

This way, an installation can auto-activate if a proper config file with key is available
inside the PsychtoolboxRoot() folder. Useful for automatic mass deployments with auto
setup.