LM: Add ability to roam outside the allowed network for a IP restricted site license.
For site licenses, we want to be able to restrict use of the license to a specific local
network of an institutions site, to allow for unlimited activations within a sites network,
but not outside.

However, students and researchers are not always working from within their institution,
but also from home or while travelling. They may roam through other networks than the
sites local and allowed network. Account for this: If the product version feature flag
"AllowNetworkRoaming" is true, then we only require being signed into the sites allowed
network while activating a machine on a license. For regular operation of an already
activated machine we try to avoid any operations which would trigger a server network
sync - and thereby would cause LA_E_IP errors for use in forbidden networks, and thereby
forced deactivation of the licensed node.

If the flag is set, we no longer do a network sync for 'AuthenticationToken' and we only
do local license validation without network or server sync during normal use. Various
help texts guide the user in proper use of the site license.

Also some other refinements to status and help messages.

This is tested with regular and DMAR IP restricted licenses, with and without the new
flag set, and with offline air-gapped activated licenses.

With this, we should be able to offer site licenses with DMAR IP network restriction.