LM: Also support license activation via user credentials instead of license key.
If a email:password combo for a user is provided, we try to authenticate a corresponding
user account. If successful, and if that account only has one license associated with it,
retrieve that license and associated license key, and activate with that license key.

Licenses can be directly associated with a user account, e.g., if bought in the shop.
They can also be indirectly associated, if they are associated with a license management
organization, and the user is a member of that organization. Typical example would be
some university, with students associated with that universities organization.