PsychVulkanCore/Linux: Add workaround for latest AMDVLK v2023.Q2.2+ bugs.
AMDVLK v2023.Q2.2 and later for Linux introduced new brokeness, in that
it no longer exposes any 16 bpc framebuffer modes, neither RGBA16 unorm,
nor RGBA16F fp16, neither HDR nor SDR. RGB10A2 unorm in SDR mode is also
broken. Strong work AMDVLK team!

As this is not fixed as of the latest v2025.Q2.1, and owners of AMD
RDNA3 and later gpu's can't downgrade their driver to a version that
is old enough to not expose the bug, we need a workaround.

Here, we detect broken driver versions, ie. everything v2023.Q2.2 and
later, and if so, replace the driver enumerated colorspace+image formats
with our own hard-coded table that is known to be supported by any and
all AMD gpu's under the buggy drivers, as long as a Linux 5.14 or later
kernel is in use, iow. as long as the system is at least Ubuntu 22.04-LTS,
which ships a Linux 5.15 kernel at a minimum.

Why does this work? Because only enumeration of supported image formats
of 10 bpc and 16 bpc kind via vkGetPhysicalDeviceSurfaceFormats2KHR() is
broken, but the driver still accepts the "missing" formats if we force
them on it.

Tested with a HDR monitor in HDR and SDR mode with AMDVLK v2023.Q3.3, the
final driver version that supports pre-RDNA gpu's, e.g., Polaris and Vega.
Can't test with real NAVI class gpu's, but this should work.