PsychLicenseHandling(): Add PsychLicenseHandling('SetupLicense').
To allow forced re-enrollment of license keys with interactive setup,
even if a key is already enrolled, or a time limited trial is active
at the moment.

Also refine setup code for Matlab + Linux license management on 64-Bit
Intel Matlab, the only Matlab that exists for Linux atm.

Activate license management for Matlab on Linux from here on.

Note: For NeuroDebian, a workable location for libLexActivator.so and
the license dat file is /usr/share/psychtoolbox-3/PsychBasic/PsychPlugins/
or a symlink from there to the real location of the files.