Screen/Linux/X11: PsychSetNominalFramerate() - Refine mode tweaking.
This is just for documentation of the math needed for this in a VRR
scenario on Linux. The code itself doesn't work on modern Linux systems,
as in "less than 15 years old", as DRM/KMS drivers do not support the
XVidMode functionality needed to set the new computed mode. A different
approach is needed on modern Linux.

In general, the XF86VidMode extension is still nominally supported, but
much of its interesting functionality is dead, as the world moved on to
RandR many ages ago...