Core,Screen,Vulkan,PsychPortAudio: Add support for teaching/education licenses.
Add infrastructure and implementations to allow to restrict and/or degrade functionality
for visual/auditory stimulation in various ways if a teaching/education license is in use on
a license enabled Psychtoolbox variant. Stuff like always showing the startup screen, and
always for at least 10 seconds with no way to escape it, and with a different startup splash
image that points out the use of a education only - no research - license. Or showing a
permanent text message at the top and bottom border of the display pointing to the edu
license. Or restricting session duration in time, or forbidding use of certain pro features,
or limiting session time more strictly to allow for only very short "sneek peak" use of
pro features. Or degradation of a visual stimulus via a visual distractor after some time,
and degradation of an auditory stimulus via mixing in of random noise after some time.

All this is parameterizable via the Product feature flag "TeachingUseOnly": If absent or
false, then standard production mode. If true, then edu license only mode, and the
associated parameter string encodes the specific versioned set of restrictions to apply.
The parameter strings format is versioned, currently prefixed with "v1: " for the first
version. This allows upgrades of the parameter string later to v2, v3, ... if this should
become neccessary.

This is a first cut at meaningful restrictions of use, so we can offer education/teaching
licenses for cheaper, and still prevent meaningful research production use.

This is the code, the new splash startup image files follow in a separate commit.