Add MacBoynToLMS to invert LMSToMacBoyn.
Added new function.  Also modified LMSTOMacBoyn so that it can return luminance as well as MB ls chromaticity if called with a new optional flag.  This would have been a better initial design, but I left the default behavior unchanged so as not to break code that calls it without the new flag.  Tested a reasonable amount but will be working more on this for use in the project and will tweak further next little while as needed.