Screen/macOS: Remove legacy IOKit fallback method for Screen('ConfigureDisplay', 'Brightness').
This doesn't work at all on Apple Silicon machines. On IntelMacs it still works for internal displays, but it relies on
long deprecated CGDisplayIOServicePort(), and it doesn't keep the Brightness slider in the display settings UI
in sync with settings, so it is inferior to the method using SPI's, which is used by default.

Let's remove it and RIP.