Screen/OSX: Fix annoying beep for fullscreen windows also on Apple Silicon Macs.
Apple Silicon Macs use the Vulkan display backend for fullscreen windows under macOS, instead of CGL on Intel
Macs, therefore a different approach is needed to avoid the beep on keypress. As Vulkan already uses a
fullscreen Cocoa NSWindow, and we omit CGL, we now also omit CGDisplayCapture, the other trigger for
keyboard focus stealing, as it seems to be not needed for Vulkan display. This leaves the keyboard focus at all
times with the original process, Octave GUI, Matlab GUI or the Terminal.app, and the problem does not happen.

Tested on Apple Silicon macOS with octave GUI/no-GUI and Matlab R2025a GUI/no-GUI.

This should do it for avoiding annoying beeps on keypress on macOS with CLI mode octave/Matlab and with
Matlab R2025a+ for the time being. Fingers crossed!