PsychStartup(): Add fallback probe sequence for GStreamer 1.26+
GStreamer 1.26 and later on MS-Windows now install into the
C:\Program Files\gstreamer\... subfolder, instead of the old path
C:\gstreamer\... This means that if the GStreamer install path
environment variable is not set for some reason, and the probe
sequence for typical locations would be used, it woul not find a
GStreamer 1.26 or later installation.

Extend the fallback probe sequence to first probe the new default
install locations in C:\Program Files\gstreamer\ for drives C, D, E,
F, G, before trying the old locations for older GStreamer versions.

This should fix startup failure with GStreamer 1.26+ if the env
variable GSTREAMER_1_0_ROOT_MSVC_X86_64 is not defined.
The ordering of first probing the new locations, then the old
ones, also makes sure that a stale directory left over from old
GStreamer installations in the old location will not throw the
process off.