Screen/Wayland: Disable colord support, unless enabled via env var.
colord support did not work great on Weston 10+ years ago, but now
it just causes a hang on every single Wayland compositor. Likely
colord is not supported under Wayland at all in the way we wanted
to use it.

Disable by default. All its code is probably a candidate for the
chopping block.